requirements.txt
playwright>=1.40.0
SQLAlchemy>=1.4
psycopg2-binary>=2.9
python-dotenv>=1.0
pgvector>=0.4.0
requests>=2.28
beautifulsoup4>=4.12
tqdm>=4.65
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=2.0.0
Pillow>=9.0.0
ftfy>=6.1.1
